Refactor to only publishing config file and add clean up commands
