+++ src/RequestInsurance/Commands/RequestInsuranceService.php
@@ -5,11 +5,11 @@ namespace Cego\RequestInsurance\Commands;
 use Exception;
 use Carbon\Carbon;
 use Nbj\Stopwatch;
-use App\RequestInsurance;
 use Illuminate\Console\Command;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Log;
 use Illuminate\Support\Facades\Config;
+use Cego\RequestInsurance\Models\RequestInsurance;
 use Cego\RequestInsurance\Exceptions\FailedToLockRequestInsurances;
 
 class RequestInsuranceService extends Command
