+++ src/RequestInsurance/Controllers/RequestInsuranceController.php
@@ -3,10 +3,10 @@
 namespace Cego\RequestInsurance\Controllers;
 
 use Exception;
-use App\RequestInsurance;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use App\Http\Controllers\Controller;
+use Cego\RequestInsurance\Models\RequestInsurance;
 
 class RequestInsuranceController extends Controller
 {
