+++ src/RequestInsurance/ViewComponents/Status.php
@@ -2,10 +2,10 @@
 
 namespace Cego\RequestInsurance\ViewComponents;
 
-use App\RequestInsurance;
 use Illuminate\View\View;
 use Illuminate\View\Component;
 use Illuminate\Contracts\View\Factory;
+use Cego\RequestInsurance\Models\RequestInsurance;
 
 class Status extends Component
 {
